+++ src/main/java/zenpad/tabheader/TabHeaderController.java
@@ -1,11 +1,10 @@
-package zenpad.controllers;
+package zenpad.tabheader;
 
 import javax.swing.JPanel;
 import javax.swing.JTabbedPane;
 
 import zenpad.models.TabHeaderModel;
 import zenpad.ui.TabManager;
-import zenpad.views.TabHeaderView;
 
 public class TabHeaderController {
     @SuppressWarnings("unused")
